requirements.txt
gradio
tavily-python
arxiv
langchain-core
langchain-openai>=0.3.9